Rename services.
